refactor: 리팩토링 및 코드 수정
