--- newspeed/src/main/java/com/example/newspeed/common/config/FilterConfig.java
@@ -1,21 +1,0 @@
-//package com.example.newspeed.common.config;
-//
-//import jakarta.servlet.*;
-//import com.example.newspeed.common.filter.LoginFilter;
-//import org.springframework.boot.web.servlet.FilterRegistrationBean;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//
-//@Configuration
-//public class FilterConfig {
-//
-//    @Bean
-//    public FilterRegistrationBean loginFilter() {
-//        FilterRegistrationBean<Filter> filterRegistrationBean = new FilterRegistrationBean<>();
-//        filterRegistrationBean.setFilter(new LoginFilter());
-//        filterRegistrationBean.setOrder(1);
-//        filterRegistrationBean.addUrlPatterns("/*");
-//
-//        return filterRegistrationBean;
-//    }
-//}
