+++ newspeed/src/main/java/com/example/newspeed/common/config/SecurityConfig.java
@@ -1,7 +1,7 @@
-package com.example.newspeed.common.jwt;
+package com.example.newspeed.common.config;
 
 
-import com.example.newspeed.common.jwt.JwtAuthenticationFilter;
+import com.example.newspeed.common.filter.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
